feat: 全面代码质量重构 + 安全加固 + 2-opt路径优化 + 测试覆盖提升至512
核心引擎:
- solver v4: 自适应搜索参数4级规模, 2-opt后处理集成, 数据预提取避免.iloc
- cost v2: calculate_green_cost拆分为7个专用函数, 完整输入校验+异常处理
- route_optimize: 新增2-opt/Or-opt/Relocate/Exchange路径后处理模块

安全加固:
- JWT生产阻断: _reject_default_key_token()拒绝旧密钥令牌
- API Key日志脱敏: get_rate_limit_key_redacted()
- rate_limit.py 限流日志密钥脱敏
- auth.py 重构: validate_jwt_key()集中校验

工程化:
- ARCHITECTURE.md 全面重构为6层架构
- pytest.ini 新增pythonpath配置
- BRANCH_STRATEGY.md 分支策略文档
- deploy/ Nginx+Caddy HTTPS部署配置
- .github/dependabot.yml 依赖自动更新

测试:
- 512测试函数, 44测试文件, 全部通过
- test_solver.py 扩展至38测试(8测试类)
- test_route_optimize.py 新增20测试
- tracking/ + integration/ + models/ 测试补齐

+++ data_types/customer.py
@@ -27,11 +27,11 @@ class Customer:
     id: int
     lat: float
     lon: float
-    demand: float = 0.0
+    demand: int = 0
     name: Optional[str] = None
-    service_time_min: float = 0.0
-    tw_earliest: Optional[float] = None
-    tw_latest: Optional[float] = None
+    service_time_min: int = 0
+    tw_earliest: Optional[int] = None
+    tw_latest: Optional[int] = None
 
     def to_dict(self) -> Dict[str, Any]:
         """转换为字典。"""
@@ -58,7 +58,7 @@ class Customer:
             id=data["id"],
             lat=data["lat"],
             lon=data["lon"],
-            demand=data.get("demand", 0.0),
+            demand=int(data.get("demand", 0)),
             name=data.get("name"),
             service_time_min=data.get("service_time_min", 0.0),
             tw_earliest=data.get("tw_earliest"),
